+++ src/main/java/org/folio/pv/controller/ErrorHandlingController.java
@@ -1,11 +1,8 @@
 package org.folio.pv.controller;
 
-import static feign.Util.UTF_8;
-
 import java.util.List;
 import java.util.stream.Collectors;
 
-import feign.FeignException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -13,6 +10,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+import org.folio.cql2pgjson.exception.CQL2PgJSONException;
 import org.folio.pv.domain.dto.Error;
 import org.folio.pv.domain.dto.Errors;
 import org.folio.pv.domain.dto.Parameter;
@@ -55,6 +53,14 @@ public class ErrorHandlingController {
     return new Error().message(exception.getMessage())
       .type(FOLIO_EXTERNAL_OR_UNDEFINED_ERROR_TYPE)
       .addParametersItem(parameter);
+  }
+
+  @ResponseBody
+  @ResponseStatus(HttpStatus.UNPROCESSABLE_ENTITY)
+  @ExceptionHandler(CQL2PgJSONException.class)
+  public Error handleException(CQL2PgJSONException exception) {
+    return new Error().message(exception.getMessage())
+      .type(FOLIO_EXTERNAL_OR_UNDEFINED_ERROR_TYPE);
   }
 
 }